extract ZstdBinding as a separate class
